+++ app/src/main/java/com/example/pokemoncards/PokemonCardsApp.kt
@@ -1,7 +1,6 @@
 package com.example.pokemoncards
 
 import android.app.Application
-import androidx.compose.material3.windowsizeclass.WindowSizeClass
 import dagger.hilt.android.HiltAndroidApp
 
 @HiltAndroidApp
@@ -11,6 +10,5 @@ class PokemonCardsApp : Application(){
     companion object {
         var isLoginSuccessful: Boolean = false
         lateinit var currentUserId: String
-        lateinit var windowsize: WindowSizeClass
     }
 }
