+++ app/src/main/java/com/example/pokemoncards/api/PokemonApi.kt
@@ -1,5 +1,6 @@
-package com.example.pokemoncards
+package com.example.pokemoncards.api
 
+import com.example.pokemoncards.data.Main
 import io.ktor.client.HttpClient
 import io.ktor.client.call.body
 import io.ktor.client.engine.android.Android
@@ -40,19 +41,19 @@ class PokemonApi {
             }
         }
 
-        suspend fun getCardById(id: String): Main? {
-
-            val response: HttpResponse =
-                client.get("https://api.pokemontcg.io/v2/cards?q=id:$id") {
-                    url{
-                    }
-                }
-            return if (response.status.value in 200..299) {
-                response.body()
-            } else{
-                null
-            }
-        }
+//        suspend fun getCardById(id: String): Main? {
+//
+//            val response: HttpResponse =
+//                client.get("https://api.pokemontcg.io/v2/cards?q=id:$id") {
+//                    url{
+//                    }
+//                }
+//            return if (response.status.value in 200..299) {
+//                response.body()
+//            } else{
+//                null
+//            }
+//        }
     }
 
 }
