+++ app/src/main/java/com/example/pokemoncards/data/User.kt
@@ -1,4 +1,4 @@
-package com.example.pokemoncards
+package com.example.pokemoncards.data
 
 data class User(
     var email: String = "",
